task: cleanup: move translation commands into sub-folder
